% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libstable4u-wrappers.R
\name{stable_pdf_and_cdf}
\alias{stable_pdf_and_cdf}
\alias{stable_pdf2}
\alias{stable_cdf2}
\title{PDF and CDF of a skew stable distribution.}
\usage{
stable_pdf2(x, pars, parametrization = 0L, tol = 1e-12)
}
\arguments{
\item{x}{Vector of points where the pdf will be evaluated.}

\item{pars}{Vector with an initial estimation of the parameters. `pars_init = c(alpha, beta, sigma, mu)`, where
* alpha: shape / stability parameter, with 0 < alpha <= 2.
* beta: skewness parameter, with -1 <= beta <= 1.
* sigma: scale parameter, with 0 < sigma.
* mu: location parameter, with mu real.
For the `2` version, one can have out of bound parameters and
an automatic adjustment will be performed to set the parameter
to the nearest bound to avoid error message and returning NA.}

\item{parametrization}{Parametrization used for the skew stable distribution, as defined by JP Nolan (1997). By default, parametrization = 0.}

\item{tol}{Relative error tolerance (precision) of the calculated values. By default, tol = 1e-12.}
}
\value{
A numeric vector.
}
\description{
Evaluate the PDF or the CDF of the skew stable distribution with parameters
pars = c(alpha, beta, sigma, mu) at the points given in x.\cr\cr
_parametrization_ argument specifies the parametrization used for the distribution
as described by JP Nolan (1997). The default value is _parametrization_ = 0.\cr\cr
_tol_ sets the relative error tolerance (precision) to _tol_. The default value is tol = 1e-12.
}
\examples{
##take out of bound pars and force them in bounds
## ***internally***
pars <- c(2.5, 0.9, 1, 0)
x <- seq(-5, 10, 0.001)

pdf <- stable_pdf2(x, pars)
cdf <- stable_cdf2(x, pars)

## following would error:
## because pars[1] (alpha > 2)
\dontrun{
pdf <- stable_pdf2(x, pars)
cdf <- stable_cdf2(x, pars)
}
}
\references{
Nolan JP (1997). Numerical Calculation of Stable Densities and Distribution Functions. Stochastic Models, 13(4) 759-774.
}
\author{
Javier Royuela del Val, Federico Simmross Wattenberg and Carlos Alberola López\cr\cr
        Maintainer: Javier Royuela del Val <jroyval@lpi.tel.uva.es>
}
\keyword{distribution}
